+++ BookingsApi/BookingsApi.Contract/V2/Requests/EditableUpdateJudiciaryParticipantRequestV2.cs
@@ -1,0 +1,10 @@
+namespace BookingsApi.Contract.V2.Requests
+{
+    public class EditableUpdateJudiciaryParticipantRequestV2 : UpdateJudiciaryParticipantRequestV2
+    {
+        /// <summary>
+        /// The participant's judicial personal code
+        /// </summary>
+        public string PersonalCode { get; set; }
+    }
+}
