+++ BookingsApi/BookingsApi.Contract/V2/Requests/Enums/JudiciaryParticipantHearingRoleCodeV2.cs
@@ -1,0 +1,8 @@
+namespace BookingsApi.Contract.V2.Requests.Enums
+{
+    public enum JudiciaryParticipantHearingRoleCodeV2
+    {
+        Judge = 1,
+        PanelMember = 2
+    }
+}
