+++ BookingsApi/BookingsApi.Contract/V2/Requests/UpdateEndpointRequestV2.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace BookingsApi.Contract.V2.Requests
+{
+    public class UpdateEndpointRequestV2 : EndpointRequestV2
+    {
+        public Guid Id { get; set; }
+    }
+}
