+++ BookingsApi/BookingsApi/Extensions/ContractExtensions.cs
@@ -1,5 +1,6 @@
 using BookingsApi.Contract.V1.Requests;
 using BookingsApi.Contract.V2.Requests;
+using BookingsApi.Contract.V2.Requests.Enums;
 namespace BookingsApi.Extensions;
 
 public static class ContractExtensions
